infrastructure for loading game. Currently we get empty level after loading.

+++ src/main/java/kniemkiewicz/jqblocks/ingame/level/LevelGenerator.java
@@ -37,7 +37,7 @@ public class LevelGenerator {
     t.logTimeAndRestart("generate objects");
   }
 
-  public void setSeed(int seed) {
+  public void setSeed(long seed) {
     random.setSeed(seed);
   }
 }
